Removes example changelog validation
The hash is the same as in the changelog on my machine, and there seems to be no difference between the file here and on Polaris in terms of actual content.

D:\GitHub\LonestarStation\html\changelogs>certutil -hashfile example.yml MD5
MD5 hash of example.yml:
88490b460c26947f5ec1ab1bb9fa9f17
CertUtil: -hashfile command completed successfully.