Big Job Update Framework
This here is just the framework for the Lonestar Jobs update. It still needs whitelisting added for restricted roles, as well as Governor-Voting & Governor-Abilities, and appropriate outfits for each added role.

However, it DOES add a number of centcom and 'antagonist' roles to the occupations section of the character screen.  Additionally, it adds several required objects to the map because the map is full of holes.